Time: 1 ms (97.55%), Space: 42.6 MB (70.89%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.java
@@ -8,6 +8,6 @@ class Solution {
         else if(mid>n/mid) hi = mid-1;
         else lo = mid+1;
     }
-    return (int)hi;
+    return hi;
     }
 }
